ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.25.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
nltk==3.8.1
python-multipart==0.0.6
